Migrate Guarantee to kernel
