+++ core/src/main/scala/tofu/package.scala
@@ -1,29 +1,10 @@
 import cats.effect.Bracket
-import cats.{ApplicativeError, MonadError}
 import cats.data.Ior
+import tofu.kernel.KernelTypes
 
-package object tofu {
-  type In[C, F[_]] = WithContext[F, C]
+package object tofu extends KernelTypes {
 
-  type HasContext[F[_], C] = Context[F] { type Ctx = C }
-
-  type HasLocal[F[_], C] = Local[F] { type Ctx = C }
-
-  type HasProvide[F[_], G[_], C] = WithProvide[F, G, C]
-
-  type HasContextRun[F[_], G[_], C] = WithRun[F, G, C]
-
-  type ApplicativeThrow[F[_]] = ApplicativeError[F, Throwable]
-  type MonadThrow[F[_]]       = MonadError[F, Throwable]
-  type BracketThrow[F[_]]     = Bracket[F, Throwable]
-
-  type Throws[F[_]]  = Raise[F, Throwable]
-  type Catches[F[_]] = Handle[F, Throwable]
-  type Tries[F[_]]   = Errors[F, Throwable]
-
-  type TConst[A, B] = A
-
-  private[tofu] type AnyK[_] = Any
+  type BracketThrow[F[_]] = Bracket[F, Throwable]
 
   type IorC[C[_], E, A] = Ior[C[E], C[A]]
