+++ kernelCE2Interop/src/main/scala/tofu/interop/CE2Kernel.scala
@@ -14,6 +14,10 @@ import scala.concurrent.duration.FiniteDuration
 import tofu.Timeout
 import tofu.internal.NonTofu
 import tofu.compat.unused
+import cats.effect.Bracket
+import tofu.Finally
+import tofu.kernel.types._
+import cats.effect.ExitCase
 
 object CE2Kernel {
   def delayViaSync[K[_]](implicit KS: Sync[K]): Delay[K] =
@@ -33,4 +37,17 @@ object CE2Kernel {
     override def timeoutTo[A](fa: F[A], after: FiniteDuration, fallback: F[A]): F[A] =
       Concurrent.timeoutTo(fa, after, fallback)
   }
+  
+  final implicit def finallyFromBracket[F[_], E](implicit F: Bracket[F, E]): Finally[F, TConst[ExitCase[E], *]] =
+    new Finally[F, TConst[ExitCase[E], *]] {
+      def finallyCase[A, B, C](init: F[A])(action: A => F[B])(release: (A, ExitCase[E]) => F[C]): F[B] =
+        F.bracketCase(init)(action) { case (a, exit) =>
+          F.void(release(a, exit))
+        }
+      def bracket[A, B, C](init: F[A])(action: A => F[B])(release: (A, Boolean) => F[C]): F[B]         =
+        F.bracketCase(init)(action) {
+          case (a, ExitCase.Completed) => F.void(release(a, true))
+          case (a, _)                  => F.void(release(a, false))
+        }
+    }
 }
